builtin/stash: factor out generic function to look up stash info
We have a function called get_stash_info that looks up this data based
on a set of command-line arguments and produces diagnostics to the user
on failure.  While this is helpful in the existing use cases, we'd like
to make use of this logic in a more programmatic way in the future.

Split out much of the function into a function which can be used
internally and which knows how to suppress these error messages with a
quiet parameter.  Wire up the rest of the function to call this internal
function to preserve the existing behavior.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>